services/db/requirements.txt
tortoise-orm==0.19.2
python-multipart==0.0.5
fastapi==0.85.1
uvicorn==0.18.3
pydantic==1.10.2